# source/app/blueprints/rest/openapi.generated.yaml
openapi: 3.1.0
info:
  title: IRIS
  version: 2.1.0
  description: "IRIS REST API \u2014 automatically generated from the Flask backend."
  contact:
    name: DFIR-IRIS
    email: contact@dfir-iris.org
  license:
    name: LGPLv3
    url: https://www.gnu.org/licenses/lgpl-3.0.txt
security:
- ApiKey: []
paths:
...
  /api/v2/manage/customers:
    get:
      operationId: search_customers_get
      summary: List customers
      tags:
      - ManageCustomers
      security:
      - ApiKey: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
                  last_page:
                    type: integer
                  current_page:
                    type: integer
                  next_page:
                    type:
                    - integer
                    - 'null'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: create_customer_post
      summary: Create a customer
      tags:
      - ManageCustomers
      security:
      - ApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v2/manage/customers/{identifier}:
    delete:
      operationId: delete_user_delete
      summary: Delete a customer
      tags:
      - ManageCustomers
      security:
      - ApiKey: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      responses:
        '204':
          $ref: '#/components/responses/Deleted'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      operationId: get_customer_get
      summary: Get a customer
      tags:
      - ManageCustomers
      security:
      - ApiKey: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: put_customer_put
      summary: Update a customer
      tags:
      - ManageCustomers
      security:
      - ApiKey: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/manage/customers/{identifier}/contacts:
    get:
      operationId: list_customer_contacts_get
      summary: List a customer contacts
      tags:
      - ManageCustomers
      security:
      - ApiKey: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: create_customer_contact_post
      summary: Create a customer contact
      tags:
      - ManageCustomers
      security:
      - ApiKey: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contact'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/manage/customers/{identifier}/contacts/{contact_id}:
    delete:
      operationId: delete_customer_contact_delete
      summary: Delete a customer contact
      tags:
      - ManageCustomers
      security:
      - ApiKey: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      - name: contact_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '204':
          $ref: '#/components/responses/Deleted'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      operationId: get_customer_contact_get
      summary: Get a customer contact
      tags:
      - ManageCustomers
      security:
      - ApiKey: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      - name: contact_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: put_customer_contact_put
      summary: Update a customer contact
      tags:
      - ManageCustomers
      security:
      - ApiKey: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      - name: contact_id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contact'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v2/manage/users/{identifier}/customers:
    put:
      operationId: put_user_customers_put
      summary: Replace user customer membership
      tags:
      - ManageUsers
      security:
      - ApiKey: []
      description: 'Body: `{customers: [int, ...]}`. The data-layer call mirrors the

        legacy `update_user_customers`; here we just normalise the body.'
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: Authorization
      description: "Bearer <api-key> \u2014 the key from your profile."
  responses:
    ApiError:
      description: Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              data: {}
            required:
            - message
    Forbidden:
      description: Insufficient permissions to access resource
    NotFound:
      description: Resource not found
    Deleted:
      description: Resource successfully deleted
  schemas:
...
    Contact:
      type: object
      properties:
        contact_name: {}
        contact_email: {}
        contact_work_phone: {}
        contact_mobile_phone: {}
        contact_role: {}
        contact_note: {}
        client_id: {}
      required:
      - contact_name
      - client_id
      description: Schema for serializing and deserializing Contact objects.
    Customer:
      type: object
      properties:
        customer_name: {}
        customer_description: {}
        customer_sla: {}
        customer_id: {}
      required:
      - customer_name
      description: Schema for serializing and deserializing Customer objects.
...